:confused::confused::confused: the crc can now be correctly read, but it confuses me, why i need to use an offset of 2 bytes in the char iterator

+++ src/chunk/chunk.h
@@ -19,7 +19,7 @@ class Chunk
 		char _type[5];
 
 		//the cyclic redundancy check number
-		int _cyclicRedundancyCheck;
+		unsigned long _cyclicRedundancyCheck;
 
 		//the position of the first letter of the Chunk (for "IDAT" its "I")
 		std::vector<unsigned char>::iterator _typePos;
